# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LaunchPad API
  version: 0.8.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /me/meetings:
    get:
      summary: List the caller's own meetings, soonest first
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
...
  /meetings:
    get:
      summary: >-
        List meetings visible to the caller (employees.read). Managers see
        their team's meetings; callers without an employee record see the
        whole organization.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [scheduled, completed, cancelled, no_show]
      responses:
        "200":
          description: OK
    post:
      summary: Schedule a meeting for an employee (meetings.manage)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingRequest"
      responses:
        "201":
          description: Created
        "400":
          description: Invalid type, start time, or duration
...
  /meetings/{meetingID}:
    patch:
      summary: Reschedule a scheduled meeting and its linked calendar event (meetings.manage)
      security:
        - bearerAuth: []
      parameters:
        - { name: meetingID, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleMeetingRequest"
      responses:
        "200": { description: Rescheduled }
        "409": { description: Meeting is not scheduled }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateMeetingRequest:
      type: object
      required: [title, type, attendeeEmployeeId, startsAt]
      properties:
        title: { type: string }
        type:
          type: string
          enum: [manager_intro, hr_orientation, team_intro, buddy_checkin, architecture_walkthrough, role_coaching, first_week_review]
        attendeeEmployeeId: { type: string }
        startsAt: { type: string, format: date-time }
        durationMin: { type: integer, minimum: 5, maximum: 480, default: 30 }
        location: { type: string, description: Meeting URL or free-text location }
        notesLink: { type: string }
...
    RescheduleMeetingRequest:
      type: object
      required: [startsAt, durationMin]
      properties:
        startsAt: { type: string, format: date-time }
        durationMin: { type: integer, minimum: 5, maximum: 480 }
        location: { type: string }
...